requirements.txt
pandas>=2.2.0,<3.0.0
psycopg2-binary>=2.9.9,<3.0
python-dotenv>=1.0.1,<2.0
requests>=2.31.0,<3.0
dbt-postgres==1.7.6